requirements.txt
asgiref==3.7.2
click==8.1.7
cloudinary==1.36.0
colorama==0.4.6
cssbeautifier==1.14.9
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.22
django-allauth==0.41.0
django-contrib-comments==2.2.0
django-countries==7.2.1
django-crispy-forms==1.14.0
djlint==1.34.0
EditorConfig==0.12.3
gunicorn==21.2.0
html-tag-names==0.1.2
html-void-elements==0.1.0
jsbeautifier==1.14.9
oauthlib==3.2.2
pathspec==0.11.2
Pillow==10.1.0
psycopg2==2.9.9
python3-openid==3.2.0
pytz==2023.3.post1
regex==2023.10.3
requests-oauthlib==1.3.1
sqlparse==0.4.4
stripe==7.4.0
tqdm==4.66.1
urllib3==1.26.15
whitenoise==6.6.0
